Fri, Jul 28, 2023, 4:09 PM +06:00

+++ beecrowd/1255.c
@@ -18,8 +18,8 @@ int main(void)
         int j = 0;
         while(text[j] != '\0')
         {
-            if((text[j] >= 'A') && (text[j] <= 'Z')) freq[(int)text[j] + 32]++;
-            else freq[(int)text[j]]++;
+            if((text[j] >= 'A') && (text[j] <= 'Z')) freq[text[j] + 32]++;
+            else freq[text[j]]++;
             j++;
         }
